refactor tensor properties
